refactor: promote local memory orchestration to services (#4901)
