+++ server/services/loraTraining/memoryPrep.js
@@ -1,8 +1,8 @@
 /** Training-specific memory policy built on the shared local-memory preflight. */
 
-import { prepareLocalMemory } from '../../lib/localMemory.js';
+import { prepareLocalMemory } from '../localMemory.js';
 
-export { unloadResidentModels, getAvailableMemoryGb, isLocalBackendUrl, gpuBlockersMessage } from '../../lib/localMemory.js';
+export { unloadResidentModels, getAvailableMemoryGb, isLocalBackendUrl, gpuBlockersMessage } from '../localMemory.js';
 
 // Below this floor even a small 4-bit training run risks swap-thrash. Larger
 // run sizing remains training-specific and still consumes `budgetGb` below.
